+++ Billedebehandling - Teis/Billedebehandling - projekt/Behandling.h
@@ -6,12 +6,12 @@ class Behandling
 {
 public:
 	Behandling();
-	int goertzler(int fs, int f, std::vector<float> sampels);
-	int bestLowDTMF(int fs, std::vector<float>sampels);
-	int bestHighDTMF(int fs, std::vector<float>sampels);
+	int goertzler(int fs, int f, std::vector<int> sampels);
+	int bestLowDTMF(int fs, std::vector<int>sampels);
+	int bestHighDTMF(int fs, std::vector<int>sampels);
 	~Behandling();
 protected:
 	int samplingFrekvens;
-	std::vector<float> maalinger;
+	std::vector<int> maalinger;
 };
